% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consecutive_median.R
\name{consecutive_median}
\alias{consecutive_median}
\title{Create seasonal median for targets requiring consecutive years}
\usage{
consecutive_median(df, n_years = 2)
}
\arguments{
\item{df}{A dataframe with values provided by which_target_df()}

\item{n_years}{The number of consecutive years over which to calculate the seasonal median.}
}
\value{
The dataframe consisting of calculated seasonal medians
}
\description{
Takes a dataframe from which_target_df(), calculates the consecutive year seasonal median results, and returns the dataframe with only
seasonal median results
}
\examples{
consecutive_year_seasonal_median(df, n_years = 2)
}
