% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LU_ben_use.R
\docType{data}
\name{LU_ben_use}
\alias{LU_ben_use}
\title{Beneficial use IDs}
\format{
An object of class \code{data.frame} with 14 rows and 2 columns.
}
\usage{
data(LU_ben_use)
}
\description{
Look up table of the beneficial use IDs and corresponding beneficial uses.
}
\details{
Data fields include:
\itemize{
  \item ben_use_id: The ID value for the beneficial use.
  \item ben_use: The beneficial use name.
}
}
\examples{
LU_ben_use1 <- data(LU_ben_use)
LU_ben_use2 <- odeqtmdl::LU_ben_use

}
\keyword{Oregon}
\keyword{beneficial}
\keyword{datasets}
\keyword{uses}
