% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LU_ben_use_code.R
\docType{data}
\name{LU_ben_use_code}
\alias{LU_ben_use_code}
\title{Beneficial use codes and corresponding beneficial use IDs}
\format{
An object of class \code{data.frame} with 148 rows and 3 columns.
}
\usage{
data(LU_ben_use_code)
}
\description{
Look up table of the applicable beneficial use IDs using the beneficial use code.
Each assessment unit has a beneficial use code assigned. This table cane be used
to determine which beneficial uses have been designated on which assessment units.
}
\details{
Data fields include:
\itemize{
  \item ben_use_code: Code assigned to the unique group of designated beneficial uses.
  \item ben_use_id: The ID value for the beneficial use.
  \item ben_use: The beneficial use name.
}
}
\examples{
LU_ben_use_code1 <- data(LU_ben_use_code)
LU_ben_use_code2 <- odeqtmdl::LU_ben_use_code

}
\keyword{Oregon}
\keyword{beneficial}
\keyword{datasets}
\keyword{uses}
