% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LU_pollutant.R
\docType{data}
\name{LU_pollutant}
\alias{LU_pollutant}
\title{Pollutant parameter look up}
\format{
An object of class \code{data.frame} with 191 rows and 7 columns.
}
\usage{
data(LU_pollutant)
}
\description{
Lookup table of DEQ and EPA ATTAINS water quality parameter names and IDs.
}
\details{
Data fields include:
\itemize{
  \item Pollu_ID: DEQ water quality parameter ID.
  \item Pollutant_DEQ: DEQ water quality parameter name.
  \item CAS: Chemical Abstracts Service registry number.
  \item WQS_Unit: Unit used for the parameter in the water quality standards.
  \item Unit_UID: Unit ID.
  \item Attains_Pollutant: EPA ATTAINS water quality parameter name.
  \item Attains_Group: EPA ATTAINS water quality parameter group.
}
}
\examples{
polluLU1 <- data(LU_pollutant)
polluLU2 <- odeqtmdl::LU_pollutant

}
\keyword{Oregon}
\keyword{TMDL}
\keyword{database}
\keyword{datasets}
