% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LU_wqstd.R
\docType{data}
\name{LU_wqstd}
\alias{LU_wqstd}
\title{Pollutant parameter and Water quality standards code look up}
\format{
An object of class \code{data.frame} with 175 rows and 2 columns.
}
\usage{
data(LU_wqstd)
}
\description{
Look up water quality standard codes by pollutant parameter ID.
}
\details{
Data fields include:
\itemize{
  \item Pollu_ID: DEQ water quality parameter ID.
  \item wqstd_code: Code assigned to each four-digit Oregon water quality standards rule number in OAR 340-41.
}
}
\examples{
wqstd_LU1 <- data(LU_wqstd)
wqstd_LU2 <- odeqtmdl::LU_wqstd

}
\keyword{Oregon}
\keyword{datasets}
\keyword{quality}
\keyword{standards}
\keyword{water}
