% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LU_wqstd_code.R
\docType{data}
\name{LU_wqstd_code}
\alias{LU_wqstd_code}
\title{Water quality standards codes}
\format{
An object of class \code{data.frame} with 17 rows and 2 columns.
}
\usage{
data(LU_wqstd_code)
}
\description{
Look up table of Oregon DEQ water quality standard codes.
}
\details{
Data fields include:
\itemize{
  \item wqstd_code: Code assigned to each four-digit Oregon water quality standards rule number in OAR 340-41.
  \item wqstd: Oregon water quality standard name.
}
}
\examples{
wqstd_code_LU1 <- data(LU_wqstd_code)
wqstd_code_LU2 <- odeqtmdl::LU_wqstd_code

}
\keyword{Oregon}
\keyword{datasets}
\keyword{quality}
\keyword{standards}
\keyword{water}
