% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_mean.R
\name{monthly_mean}
\alias{monthly_mean}
\title{Create monthly mean values for a dataset}
\usage{
monthly_mean(df)
}
\arguments{
\item{df}{A dataframe with values provided by which_target_df()}
}
\value{
The dataframe consisting of calculated monthly means
}
\description{
Takes a dataframe from which_target_df(), calculates the monthly mean results, and returns the dataframe with only
monthly mean results
}
\examples{
monthly_mean(df)
}
