% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_median.R
\name{monthly_median}
\alias{monthly_median}
\title{Create monthly median values for a dataset}
\usage{
monthly_median(df)
}
\arguments{
\item{df}{A dataframe with values provided by which_target_df()}
}
\value{
The dataframe consisting of calculated monthly medians
}
\description{
Takes a dataframe from which_target_df(), calculates the monthly median results, and returns the dataframe with only
monthly median results
}
\examples{
monthly_median(df)
}
