% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_mean.R
\name{seasonal_mean}
\alias{seasonal_mean}
\title{Create seasonal mean values for a dataset}
\usage{
seasonal_mean(df)
}
\arguments{
\item{df}{A dataframe with values provided by which_target_df()}
}
\value{
The dataframe consisting of calculated seasonal means
}
\description{
Takes a dataframe from which_target_df(), calculates the seasonal mean results, and returns the dataframe with only
seasonal mean results
}
\examples{
seasonal_mean(df)
}
