% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonal_median.R
\name{seasonal_median}
\alias{seasonal_median}
\title{Create seasonal median values for a dataset}
\usage{
seasonal_median(df)
}
\arguments{
\item{df}{A dataframe with values provided by which_target_df()}
}
\value{
The dataframe consisting of calculated seasonal medians
}
\description{
Takes a dataframe from which_target_df(), calculates the seasonal median results, and returns the dataframe with only
seasonal median results
}
\examples{
seasonal_median(df)
}
