% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_assessment.R
\name{target_assessment}
\alias{target_assessment}
\title{Create target statistical base values for a dataset}
\usage{
target_assessment(df)
}
\arguments{
\item{df}{A dataframe with values provided by which_target_df()}
}
\value{
The dataframe consisting of calculated target statistical bases
}
\description{
Takes a dataframe from which_target_df(), calculates the target statistical base results, and returns the dataframe with only
target statistical base results
}
\examples{
target_assessment(df)
}
