% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmdl_actions.R
\docType{data}
\name{tmdl_actions}
\alias{tmdl_actions}
\title{Oregon TMDL Actions}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 55 rows and 10 columns.
}
\usage{
data(tmdl_actions)
}
\description{
Summary of all non tribal TMDL actions applicable in Oregon
}
\details{
\itemize{
  \item action_id:	USEPA ATTAINS Action ID assigned to each TMDL document.
  \item TMDL_name:	Name of TMDL document.
  \item TMDL_issue_year:	Year the TMDL was issued by the issue agency..
  \item TMDL_active:	Boolean to indicate if the TMDL and TMDL allocations are effective and being implemented.
  \item issue_agency:	The agency that developed and issued the TMDL (OR or EPA).
  \item in_attains:	Boolean to indicate if the TMDL action has been entered into EPA's ATTAINS database.
  \item TMDL_issue_date:	The date the TMDL was issued by the issue agency.
}
}
\examples{
actions1 <- data(tmdl_actions)
actions2 <-odeqtmdl::tmdl_actions

}
\keyword{Oregon}
\keyword{TMDL}
\keyword{database}
\keyword{datasets}
