% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmdl_db.R
\docType{data}
\name{tmdl_db}
\alias{tmdl_db}
\title{Oregon TMDL database}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 159171 rows and 23 columns.
}
\usage{
data(tmdl_db)
}
\description{
Oregon TMDL database
}
\examples{
db1 <- data(tmdl_db)
db2 <-odeqtmdl::tmdl_db

}
\keyword{Oregon}
\keyword{TMDL}
\keyword{database}
\keyword{datasets}
