% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmdl_export_xlsx.R
\name{tmdl_export_xlsx}
\alias{tmdl_export_xlsx}
\title{Export TMDL database info to an excel spreadsheet.}
\usage{
tmdl_export_xlsx(action_ids = NULL, xlsx_filename = NULL)
}
\arguments{
\item{action_ids}{vector of TMDL action IDs to export. Required.}

\item{xlsx_filename}{The path and file name of the output excel file.}
}
\description{
Exports all or a subset of the Oregon TMDL database including
action table (\code{\link{tmdl_actions}}),
assessment unit table (\code{\link{tmdl_au}}),
beneficial use table (\code{\link{tmdl_ben_use}}),
geo_id table (\code{\link{tmdl_geo_id}}),
TMDL target table (\code{\link{tmdl_targets}}),
and NPDES permittee WLA table (\code{\link{tmdl_wla}}) as an excel workbook.
The workbook format is based on DEQ TMDL database template version 3.0.
}
