% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmdl_geo_ids.R
\docType{data}
\name{tmdl_geo_ids}
\alias{tmdl_geo_ids}
\title{Oregon TMDL geo IDs}
\format{
An object of class \code{data.frame} with 111 rows and 4 columns.
}
\usage{
data(tmdl_geo_ids)
}
\description{
Inventory of unique TMDL geo IDs for non-tribal Oregon TMDLs. A geo ID is a
unique ID that is used to identify where a specific TMDL target applies. Geo IDs
are mapped to the National Hydropgrahy Dataset version NHDH_OR_931v220.
Note the inventory is incomplete. See each relevant TMDL document
for more information on where the TMDL applies. A full listing of all
non-tribal TMDL actions in Oregon can be viewed using \code{\link{tmdl_actions}}.
}
\details{
\itemize{
  \item action_id:	EPA ATTAINS Action ID assigned to each TMDL document.
  \item geo_id:	Unique ID assigned to the NHD reaches codes where a TMDL target applies. ID is structured as YearTMDLissued_ShortTMDLdocName_TargetGeoArea.
  \item geo_description:	General name and description of area where the TMDL target applies.
  \item mapped:	Boolean to indicate if the geo_id has been mapped in a GIS.
}
}
\examples{
db1 <- data(tmdl_geo_ids)
db2 <-odeqtmdl::tmdl_geo_ids

}
\keyword{Oregon}
\keyword{TMDL}
\keyword{database}
\keyword{datasets}
