% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmdl_update.R
\name{tmdl_update}
\alias{tmdl_update}
\title{Update TMDL tables with new or revised information}
\usage{
tmdl_update(
  action_ids = NULL,
  xlsx_template,
  gis_path,
  package_path,
  update_tables = TRUE,
  update_reaches = TRUE
)
}
\arguments{
\item{action_ids}{vector of TMDL action IDs to update. Required.}

\item{xlsx_template}{The path including file name of the TMDL excel template.}

\item{gis_path}{Path to the directory that holds the TMDL
GIS shapefiles. The GIS shapefiles can be in a subdirctory. The function
completes a recursive search using \code{\link[base]{list.files}}. The GIS files
must be ESRI shapefiles and named in the following way:
    \itemize{
      \item NHD reaches: 'action_[action_id]_[parameter]_[pollutant]_NHD'
      \item AU Flowlines: 'action_[action_id]_[parameter]_[pollutant]_AU_Flowline'
      \item AU Waterbodies: 'action_[action_id]_[parameter]_[pollutant]_AU_WB'
      \item Geo IDs: 'action_[action_id]_[geo_id]_geo_ids'
      }
      where, action_id, parameter, and pollutant are user supplied.}

\item{package_path}{Path to the top level directory of the odeqtmdl R package. The 'data', data_raw', and 'inst/extdata' folders must exist.}

\item{update_tables}{logical. If TRUE, updates the following package tables:
\itemize{
  \item tmdl_actions
  \item tmdl_parameters
  \item tmdl_geo_id
  \item tmdl_targets
  \item tmdl_wla
  \item tmdl_ben_use
  }}

\item{update_reaches}{logical. if TRUE, imports GIS files and updates the following package tables:
\itemize{
  \item tmdl_target_reaches
  \item tmdl_au
  \item tmdl_au_gnis
  }}
}
\description{
This function is intended to be used by DEQ staff only.
Imports TMDL information from DEQ's TMDL database template and GIS shapefiles; and
add the new or revised the information to the odeqtmdl package tables. To
run this function the xlsx template and/or the parameter/pollutant pair GIS files
must be completed.
}
\details{
Updates are made to the existing package tables by first removing all rows
with the same 'action_id' identified in \code{action_ids}. The new
information from the \code{xlsx_template} and any GIS files are then added to the package tables. There
is no filtering for specific parameters or pollutants. All
information attributed to the action_id is removed before being replaced, so
all information attributed to the action must be included in the \code{xlsx_template}.
}
