% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmdl_wqstd.R
\docType{data}
\name{tmdl_wqstd}
\alias{tmdl_wqstd}
\title{TMDL pollutant parameter and Water quality standard table}
\format{
An object of class \code{data.frame} with 3764 rows and 3 columns.
}
\usage{
data(tmdl_wqstd)
}
\description{
Look up water quality standard codes by pollutant parameter ID. This information is used to determine which water quality standard was addressed by the TMDL.
}
\details{
Data fields include:
\itemize{
  \item action_id:	USEPA ATTAINS Action ID assigned to each TMDL document.
  \item Pollu_ID: DEQ water quality parameter ID. ID is for the parameter 'TMDL_wq_limited_parameter' in the TMDL tables.
  \item wqstd_code: Code assigned to each four-digit Oregon water quality standards rule number in OAR 340-41.
}
}
\examples{
tmdl_wqstd_LU1 <- data(tmdl_wqstd)
tmdl_wqstd_LU2 <- odeqtmdl::tmdl_wqstd

}
\keyword{Oregon}
\keyword{datasets}
\keyword{quality}
\keyword{standards}
\keyword{water}
