% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_target.R
\name{which_target}
\alias{which_target}
\title{Which TMDL target geio ID applies}
\usage{
which_target(reachcode, parameter, date)
}
\arguments{
\item{reachcode}{The NHD reachcode for the monitoring location}

\item{parameter}{The AWQMS pollutant or characteristic name value for the sample}

\item{date}{The sample date in yyyy-mm-dd format}
}
\value{
The applicable TMDL target
}
\description{
Takes a sample reachcode, parameter, and date and returns the TMDL target ID(s) that apply
}
\examples{
which_target(reachcode = "17090012000099", parameter = "Total suspended solids", date = "2015-07-01")
}
