% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_target_df.R
\name{which_target_df}
\alias{which_target_df}
\title{Which TMDL target ID applies}
\usage{
which_target_df(df, all_obs = TRUE)
}
\arguments{
\item{df}{Data frame in which to add target information}
}
\value{
Data frame with tmdl target columns
}
\description{
Takes a dataframe with sample reachcode, parameter, and date and returns the TMDL target info
}
\examples{
which_target(df)
}
