% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odeqtmdl-package.R
\docType{package}
\name{odeqtmdl-package}
\alias{odeqtmdl}
\alias{odeqtmdl-package}
\title{odeqtmdl: Oregon TMDL database}
\description{
The odeqtmdl R package includes databases of TMDL information from non-tribal Oregon TMDLs and a set of functions that assist in assessment of water quality data against select TMDL targets. Currently the TMDL databases are incomplete. See each relevant TMDL document for the official record.
}
\author{
\strong{Maintainer}: Ryan Michie \email{Ryan.Michie@deq.oregon.gov}

Authors:
\itemize{
  \item Colin Donald \email{Colin.Donald@deq.oregon.gov}
}

}
\keyword{internal}
