% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmdl_actions.R
\docType{data}
\name{tmdl_actions}
\alias{tmdl_actions}
\title{Oregon TMDL Actions}
\format{
An object of class \code{data.frame} with 70 rows and 12 columns.
}
\usage{
data(tmdl_actions)
}
\description{
Summary of TMDL actions applicable in Oregon. TMDLs developed by tribal governments are not included at this time.
}
\details{
** Field removed in version 0.6.
See \code{\link{tmdl_parameters}} for listings of all unique TMDL parameter and pollutant pairs.

Database fields include:
\itemize{
  \item action_id:	USEPA ATTAINS Action ID assigned to each TMDL document.
  \item TMDL_name:	Name of TMDL document.
  \item TMDL_issue_year:	Year the TMDL was issued by the issue agency.
  \item issue_agency:	The agency that developed and issued the TMDL (Oregon DEQ, IDEQ, WADOE, or EPA).
  \item in_attains:	logical to indicate if the TMDL action has been entered into EPA's ATTAINS database.
  \item attains_status: Status of TMDL action in ATTAINS. NA indicates the TMDL action is not included in ATTAINS.
  \item TMDL_issue_date:	The date the TMDL was issued by the issue agency.
  \item EPA_action_date: The date EPA took action (approval or disapproval) on the TMDL.
  \item citation_abbreviated: Abbreviated citation of TMDL document using DEQ style guidelines (Chicago Manual of Style).
  \item citation_full: Full citation of TMDL document using DEQ style guidelines (Chicago Manual of Style).
  \item TMDL_comment: Note summarizing high level information about the TMDL overall.
  \item URL: Website address for the TMDL.

}
}
\examples{
actions1 <- data(tmdl_actions)
actions2 <- odeqtmdl::tmdl_actions

}
\keyword{Oregon}
\keyword{TMDL}
\keyword{database}
\keyword{datasets}
