% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmdl_aus.R
\docType{data}
\name{tmdl_aus}
\alias{tmdl_aus}
\title{Oregon TMDL assessment unit database}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16843 rows and 25 columns.
}
\usage{
data(tmdl_aus)
}
\description{
Inventory of assessment units where non-tribal Oregon TMDLs have been developed. Note the inventory
is still being developed and some information may not be accurate.
See each relevant TMDL document for more information and applicability. A full listing of all
non-tribal TMDL actions in Oregon can be viewed using \code{\link{tmdl_actions}}.
}
\details{
NHD values are derived from NHDH_OR_931v220, which is the current version used for DEQ business data. Database fields include:

\itemize{
  \item action_id:	EPA ATTAINS Action ID assigned to each TMDL document.
  \item TMDL_name:	Name of TMDL document.
  \item TMDL_issue_year:	Year the TMDL was issued by the issue agency.
  \item TMDL_wq_limited_parameters:	Name of the water quality limited 303(d) parameter that the TMDL addresses.
  \item TMDL_pollutant:	Name of TMDL pollutant causing the water quality impairment.
  \item TMDL_scope: Provides information about how the TMDL applies.
     \itemize{
     \item TMDL:	Identifies segments that a TMDL was developed for.
     \item Allocation only: Identifies segments where a TMDL allocation applies
                           but the TMDL does not address a 303(d) listing in that segment.
                           Typically this situation is applicable for tributaries or canals
                           that are upstream of the segment where the "TMDL" applies.
                           The pollutant reduction in the upstream segment is needed to achieve the
                           TMDL loading capacity of the downstream segment.
                           This is common for TMDLs that address narrative water quality standards
                           or cases when the pollutant is not the same as the listed parameter (e.g. nutrients).
     \item Advisory allocation: Identifies segments that have suggested non regulatory allocations;
                               or segments that may be used to assess progress or status of allocation attainment
                               but the segment is not the regulatory compliance point as defined in the TMDL.
                               }
  \item TMDL_active: Boolean to indicate if the TMDL and TMDL allocations are effective and being implemented.
  \item Period: Identifies the fish use period that the TMDL addresses. Only used for TMDLs for temperature or dissolved oxygen.
     \itemize{
     \item year_round: TMDL developed to address only non spawning (year round) uses for temperature or dissolved oxygen water quality standards.
     \item spawning: TMDL developed to address only spawning uses for the temperature or dissolved oxygen water quality standards.
     \item Both: TMDL developed to address both spawning and non spawning (year round) uses for temperature or dissolved oxygen water quality standards.
     }
  \item citation_abbreviated: Abbreviated citation of TMDL document using DEQ style guidelines (Chicago Manual of Style).
  \item citation_full: Full citation of TMDL document using DEQ style guidelines (Chicago Manual of Style).
  \item HUC_6: Basin six digit USGS hydrological unit code
  \item HU_6_NAME: USGS Basin name.
  \item HUC6_full: Concatenation of the HUC_6 and HU_6_NAME fields.
  \item HUC_8: Subbasin six digit USGS hydrological unit code.
  \item HU_8_NAME: USGS Subbasin name.
  \item HUC8_full: Concatenation of the HUC_8 and HU_8_NAME fields.
  \item AU_ID:	Assessment Unit ID.
  \item AU_Name: Name of the assessment unit
  \item AU_Description: Assessment unit descriptions
  \item TMDL_length_km: Length of the assessment unit in kilometers where the TMDL is applicable. Length is calculated using the linear flowline feature in Albers Equal Area projection.
  \item AU_length_km: Length of the entire assessment unit in kilometers. Length is calculated using the linear flowline feature in Albers Equal Area projection.
  \item TMDL_AU_Percent: Percent of the assessment unit the TMDL has addressed.
}
}
\examples{
db1 <- data(tmdl_aus)
db2 <-odeqtmdl::tmd_aus

}
\keyword{Oregon}
\keyword{TMDL}
\keyword{assessment}
\keyword{database}
\keyword{datasets}
\keyword{unit}
