% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmdl_geo_ids.R
\docType{data}
\name{tmdl_geo_ids}
\alias{tmdl_geo_ids}
\title{Oregon TMDL geo IDs}
\format{
An object of class \code{data.frame} with 376 rows and 4 columns.
}
\usage{
data(tmdl_geo_ids)
}
\description{
Inventory of unique TMDL geo IDs for Oregon TMDLs. A geo ID is a
unique ID that is used to identify where a specific TMDL target applies. Geo IDs
are mapped to the National Hydrography Dataset version NHDH_OR_931v220.
See \code{\link{tmdl_targets}} for a summary of TMDL targets.
}
\details{
Note the TMDL information is still being assembled and is undergoing review.
Some information may not be accurate. See each relevant TMDL document for the
official record and more information.


\itemize{
  \item action_id:	EPA ATTAINS Action ID assigned to each TMDL document.
  \item geo_id:	Unique ID assigned to the NHD reaches where a TMDL target applies.
  \item geo_description:	General name and description of area where the TMDL target applies.
  \item geo_id_mapped:	Boolean to indicate if the geo_id has been mapped in a GIS.
}
}
\examples{
db1 <- data(tmdl_geo_ids)
db2 <-odeqtmdl::tmdl_geo_ids

}
\keyword{Oregon}
\keyword{TMDL}
\keyword{database}
\keyword{datasets}
