% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmdl_parameters.R
\docType{data}
\name{tmdl_parameters}
\alias{tmdl_parameters}
\title{Oregon TMDL parameters}
\format{
An object of class \code{data.frame} with 685 rows and 7 columns.
}
\usage{
data(tmdl_parameters)
}
\description{
Summary and status of all unique water quality limited parameter and pollutant pair
combinations for each TMDL action.
}
\details{
Database fields include:
\itemize{
  \item action_id:	EPA ATTAINS action ID assigned to each TMDL document.
  \item TMDL_wq_limited_parameters:	Name of the water quality limited 303(d) parameter that the TMDL addresses.
  \item TMDL_pollutant:	Name of TMDL pollutant causing the water quality impairment.
  \item TMDL_status: Status of TMDL for the parameter and pollutant.
  \itemize{
       \item Active: TMDL has been approved by EPA and is active.
       \item Not Active: TMDL has been withdrawn, disapproved by EPA, and/or replaced with a newer TMDL.
       \item In Development: TMDL is in development.
  }
  \item TMDL_status_comment: Note summarizing information about the TMDL and if it was replaced or modified.
  \item scope_narrative:	Narrative summary of where the TMDL allocations apply.
  \item revision_action_id: The EPA ATTAINS action ID assigned to the TMDL revision.
  \item TMDL_mapped: logical to indicate if the TMDL parameter and pollutant scope has mapped in a GIS.
}
}
\examples{
p1 <- data(tmdl_parameters)
p2 <- odeqtmdl::tmdl_parameters

}
\keyword{Oregon}
\keyword{TMDL}
\keyword{database}
\keyword{datasets}
