% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmdl_parameters.R
\docType{data}
\name{tmdl_parameters}
\alias{tmdl_parameters}
\title{Oregon TMDL parameters}
\format{
An object of class \code{data.frame} with 369 rows and 3 columns.
}
\usage{
data(tmdl_parameters)
}
\description{
Summary of all unique water quality limited parameter and pollutant pair
combinations for each TMDL action.
}
\details{
Database fields include:
\itemize{
  \item action_id:	EPA ATTAINS action ID assigned to each TMDL document.
  \item TMDL_parameter:	Name of the water quality limited 303(d) parameter that the TMDL addresses.
  \item TMDL_pollutant:	Name of TMDL pollutant causing the water quality impairment.
}
}
\examples{
p1 <- data(tmdl_parameters)
p2 <- odeqtmdl::tmdl_parameters

}
\keyword{Oregon}
\keyword{TMDL}
\keyword{database}
\keyword{datasets}
