% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmdl_wla.R
\docType{data}
\name{tmdl_wla}
\alias{tmdl_wla}
\title{NPDES permittees assigned a TMDL WLA}
\format{
An object of class \code{data.frame} with 3620 rows and 6 columns.
}
\usage{
data(tmdl_wla)
}
\description{
Table of NPDES permittees that were assigned a wasteload allocation (WLA) in
a TMDL. The inventory is incomplete. See each relevant TMDL document
for more information.
}
\details{
\itemize{
  \item action_id:	EPA ATTAINS Action ID assigned to each TMDL document.
  \item AU_ID:	Assessment Unit ID.
  \item TMDL_pollutant:	Name of TMDL pollutant causing the water quality impairment.
  \item EPANum: EPA NPDES facility number.
  \item WQFileNum:	DEQ NPDES facility water quality file number.
  \item facility_name:	Common Name or name in the TMDL document of the NPDES permitted facility.
}
}
\examples{
db1 <- data(tmdl_wla)
db2 <-odeqtmdl::tmdl_wla

}
\keyword{Oregon}
\keyword{TMDL}
\keyword{database}
\keyword{datasets}
