% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_target.R
\name{which_target}
\alias{which_target}
\title{Which TMDL target ID applies}
\usage{
which_target(reachcode, parameter, date)
}
\arguments{
\item{reachcode}{The GeoID for the sample location}

\item{parameter}{The pollutant or characteristic name value for the sample}

\item{date}{The sample date in yyyy-mm-dd format}
}
\value{
The applicable TMDL target
}
\description{
Takes a sample reachcode, parameter, and date and returns the TMDL target ID(s) that apply
}
\examples{
which_target(GeoID = "2006_Willamette_JohnsonCreek", parameter = "Total phosphorus, mixed forms", date = "2015-07-01")
}
